requirements.txt
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.13.4
bracketool==0.1
CacheControl==0.12.14
certifi==2025.6.15
charset-normalizer==3.4.2
cloudpickle==3.1.1
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.14
decorator==5.2.1
dnspython==2.7.0
dotenv==0.9.9
executing==2.2.0
fonttools==4.58.4
future==1.0.0
grpcio==1.73.1
grpcio-tools==1.73.1
h11==0.16.0
hyperopt==0.2.7
idna==3.10
ipykernel==6.29.5
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
llvmlite==0.44.0
matplotlib==3.10.3
matplotlib-inline==0.1.7
mongoengine==0.29.1
msgpack==1.1.1
nest-asyncio==1.6.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.9
Pygments==2.19.2
pymongo==4.13.2
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.0.0
requests==2.32.4
scikit-learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
selenium==4.34.2
setuptools==80.9.0
shap==0.48.0
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
stack-data==0.6.3
statbotics==3.0.0
threadpoolctl==3.6.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
xgboost==3.0.2
